attacks/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.5
fastapi==0.104.1
Flask==3.0.0
getmac==0.9.4
h11==0.14.0
httptools==0.6.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
netifaces==0.11.0
packaging==23.2
pyaml==23.9.7
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.5.1
pydantic_core==2.14.3
pyshark==0.6
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
termcolor==2.3.0
typing_extensions==4.8.0
unicorn==2.0.1.post1
urllib3==2.0.6
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.0
